requirements.txt
Django==1.4.5
amqp==1.0.12
anyjson==0.3.3
autopep8==0.9.2
billiard==2.7.3.28
celery==3.0.19
coverage==3.6
django-celery==3.0.17
django-configurations==0.3
django-coverage==1.2.4
django-ses==0.4.1
dogapi==1.4.0
dogstatsd-python==0.2.1
kombu==2.5.11
logilab-astng==0.24.3
logilab-common==0.59.1
mock==1.0.1
pep8==1.4.6
pycrypto>=2.6
pylint==0.28.0
python-dateutil==2.1
pytz==2013b
requests==0.14.2
six==1.3.0
wsgiref==0.1.2
